requirements.txt
# TUI Framework
textual==0.47.1

# YouTube API
google-api-python-client==2.111.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0

# Configuration
pyyaml==6.0.1

# Database (async SQLite)
aiosqlite==0.19.0

# Additional utilities
python-dateutil==2.8.2

# Graphiques et visualisation
plotext==5.2.8  # Graphiques dans le terminal
